+++ backend/api/admin/mixins.py
@@ -1,0 +1,30 @@
+import csv
+
+from django.http import HttpResponse
+
+
+class ExportCsvMixin:
+
+    def export_as_csv(self, request, queryset):
+        meta = self.model._meta
+        field_names = [field for field in self.list_display]
+
+        response = HttpResponse(content_type='text/csv')
+        response['Content-Disposition'] = 'attachment; filename={}.csv'.format(meta)
+        writer = csv.writer(response)
+
+        writer.writerow(field_names)
+        for obj in queryset:
+            writer.writerow([getattr(obj, field) for field in field_names])
+
+        return response
+
+    export_as_csv.short_description = '輸出成 csv 檔'
+
+
+class RestoreMixin:
+
+    def restore(self, request, queryset):
+        queryset.undelete()
+
+    restore.short_description = '回復'
